improvement(metrics): add namespace to pipelinerun_count metric
The pipelinerun_count metrics provides a view of how many pipelines
are being executed in a cluster. However, the metrics is missing the
namespace tag that is being added to most of the other Tekton metrics.
This means that we can't see how many pipelines are being executed for a
particular namespace in the same way we can with the other Tekton
metrics. This adds the namespace tag to the pipelinerun_count metric so
that its consistent with the other Tekton metrics.